requirements.txt
Flask==2.3.2
flask-cors==4.0.0
requests==2.31.0
transformers==4.40.0
torch==2.3.0
beautifulsoup4==4.12.3
lxml==4.9.3
python-dotenv==1.0.1
mysql-connector-python==8.3.0
